Update IEmailProvider
Accommodates the attachments parameter. The param is specified as a default of null. This works, allowing the implementation to omit this, and it will default to null. Do not give the parameter an option of null in the implementation though, or it will always be null even if , and it will work.