Update SendGridEmailProvider.cs
Basically just a compatibility change. The IEmailProvider interface wants the attachment parameter. I added it to the SendEmailAsync method for SendGrid, but it is not fully implemented. Also, since the interface sets a fallback to null, omitting it in calling code is allowed, and specifying attachments will have no effect. This might need some discussion. It is unclear to me whether Sendgrid discourages attachments anyway, so maybe it's better as is. Alternatively, the SendGrid methods could take up the attachment feature if desired.